Add fabric mod initializer, register emojis on forge
