+++ Fabric/src/main/java/com/hrznstudio/emojiful/EmojifulFabric.java
@@ -1,11 +1,10 @@
 package com.hrznstudio.emojiful;
 
-import net.fabricmc.api.ClientModInitializer;
+import net.fabricmc.api.ModInitializer;
 
-public class EmojifulFabric implements ClientModInitializer {
-
+public class EmojifulFabric implements ModInitializer {
     @Override
-    public void onInitializeClient() {
-        ClientProxy.setup();
+    public void onInitialize() {
+        //Register config, serializer and recipetype, handle recipetype reloading too
     }
 }
