+++ Fabric/src/main/java/com/hrznstudio/emojiful/EmojifulFabricClient.java
@@ -1,0 +1,11 @@
+package com.hrznstudio.emojiful;
+
+import net.fabricmc.api.ClientModInitializer;
+
+public class EmojifulFabricClient implements ClientModInitializer {
+
+    @Override
+    public void onInitializeClient() {
+        ClientProxy.setup();
+    }
+}
